Split container/direct.go into individual files
